requirements.txt
channels
django
django-reactor
pypiwin32
